Split drawing.ts into minimap.ts, worldmap.ts and traces.ts as requested.
