+++ mettagrid/player/src/common.ts
@@ -1,7 +1,6 @@
 import { Vec2f } from './vector_math.js';
 import { Context3d, parseHtmlColor } from './context3d.js';
 import { PanelInfo } from './panels.js';
-import { onFrame } from './main.js';
 
 // The 3d context, used for nearly everything.
 export const ctx = new Context3d(document.getElementById('global-canvas') as HTMLCanvasElement);
